Refactor the subscription prompt to accommodate new US elec offer
First pass at US elections prompt for testing, css work left.

- Change how the subscription offer prompt is inited so that we
can pass in flags.
- Moved some generic logic about if and which offer to show into
index.
- Split Lionel slider and US election slider into seperate files
- refactor tests
  - add tests for index, lionel and new US offer
