+++ subscription-offer-prompt/test/index.spec.js
@@ -1,114 +1,68 @@
-/* globals should, sinon */
-import SlidingPopup from 'n-sliding-popup';
-import Superstore from 'superstore';
-import { init } from '../index';
+/* globals sinon */
+import subscriptionOfferPrompt from '../index';
+import * as lionel from '../lionel';
+import * as election from '../us-elections';
 
-const delay = (ms, value) => new Promise(resolve => setTimeout(resolve.bind(null, value), ms));
+describe('Subscription Offer Prompt Init', () => {
 
-describe('Subscription Offer Prompt', () => {
-
-	const localStorage = new Superstore('local', 'n-ui.subscription-offer-prompt')
-	const sessionStorage = new Superstore('session', 'next.product-selector')
+	let lionelStub;
+	let electionStub;
+	let flags;
 
 	beforeEach(() => {
 		Object.defineProperty(document, 'cookie', { value: '', configurable: true });
-		const fetchStub = sinon.stub(window, 'fetch');
-		fetchStub
-			.withArgs('/country')
-			.returns(Promise.resolve({
-				json: () => Promise.resolve('GBR')
-			}));
-		fetchStub
-			.withArgs('https://www.howsmyssl.com/a/check')
-			.returns(Promise.resolve({
-				json: () => Promise.resolve({ tls_version: 'TLS 1.2' })
-			}));
-
-		return Promise.all([
-			localStorage.set('last-closed', Date.now() - (1000 * 60 * 60 * 24 * 30)),
-			sessionStorage.set('last-seen', Date.now())
-		])
+		lionelStub = sinon.spy(lionel, 'init');
+		electionStub = sinon.spy(election, 'init');
+		// stub out the flag.get()
+		flags = { get: () => true }
 	});
 
 	afterEach(() => {
-		window.fetch.restore();
 		delete document.cookie;
-
-		return Promise.all([
-			localStorage.unset('last-closed'),
-			sessionStorage.unset('last-seen')
-		]);
+		lionelStub.restore();
+		electionStub.restore();
+		flags = null;
 	});
 
-	it('should show prompt if navigated from barrier page, not on a barrier page, not logged in and hasnt been shown in 30 days', () =>
-		init().then(popup => popup.should.be.an instanceof(SlidingPopup))
-	);
-
-	it('should have correct attributes', () =>
-		init().then(popup => {
-			popup.el.getAttribute('class').should.include('n-sliding-popup subscription-prompt');
-			popup.el.getAttribute('data-n-component').should.equal('o-sliding-popup');
-			popup.el.getAttribute('data-n-sliding-popup-position').should.equal('bottom left');
-		})
-	);
-
-	it('should have correct html', () =>
-		init().then(popup => {
-			popup.el.innerHTML.should.contain('You qualify for a 25% subscription discount')
-		})
-	);
-
-	it('should set onClose to function', () =>
-		init().then(popup => {
-			popup.el.onClose.should.be.a('function')
-		})
-	);
-
-	it('should store date in local storage when closed', () =>
-		init()
-			.then(popup => {
-				popup.el.onClose();
-				return localStorage.get('last-closed');
-			})
-			// give a 1s buffer
-			.then(lastClosed => lastClosed.should.be.closeTo(Date.now(), 1000))
-	);
-
-	// TODO: naughty, but errors for unknown reason - https://circleci.com/gh/Financial-Times/n-ui/2829
-	xit('should ‘pop-up’ after 2 seconds', () =>
-		init()
-			.then(popup => {
-				sinon.spy(popup, 'open');
-				popup.open.should.not.have.been.called;
-				return delay(2050, popup);
-			})
-			.then(popup => popup.open.should.have.callCount(1))
-	);
-
-	it('should not not show on barrier pages', () => {
+	it('should not init either prompt if on barrier pages', () => {
 		const barrier = document.createElement('div');
 		barrier.className = 'ft-subscription-panel';
 		document.body.appendChild(barrier);
-		return init()
-			.then(popup => {
-				document.body.removeChild(barrier);
-				should.not.exist(popup);
-			});
+
+		subscriptionOfferPrompt(flags);
+		lionelStub.should.not.have.been.called;
+		electionStub.should.not.have.been.called;
+
+		document.body.removeChild(barrier);
 	});
 
-	it('should not show if logged in', () => {
+	it('should not init either prompt if logged in', () => {
 		Object.defineProperty(document, 'cookie', { value: 'FTSession=foo', configurable: true });
-		return init().then(popup => should.not.exist(popup));
+
+		subscriptionOfferPrompt(flags);
+		lionelStub.should.not.have.been.called;
+		electionStub.should.not.have.been.called;
 	});
 
-	it('should not show if last shown within 30 days', () => {
-		localStorage.set('last-closed', Date.now() + (1000 * 60 * 60 * 24 * 29));
-		return init().then(popup => should.not.exist(popup));
+	it('should not init either prompt if b2cMessagePrompt flag is false', () => {
+		flags.get = () => false;
+
+		subscriptionOfferPrompt(flags);
+		lionelStub.should.not.have.been.called;
+		electionStub.should.not.have.been.called;
 	});
 
-	it('should not show barrier page has not been visited in this session', () => {
-		sessionStorage.unset('last-seen');
-		return init().then(popup => should.not.exist(popup));
+	it('should init "Lionel slider" if NOT logged in & NOT on barrier page & NOT on /us-election-2016 page', () => {
+		subscriptionOfferPrompt(flags);
+		lionelStub.should.have.callCount(1);
+		electionStub.should.have.not.been.called;
+	});
+
+	it('should init "2016 US Election slider" if NOT logged in & on /us-election-2016 stream page', () => {
+		document.documentElement.setAttribute('data-concept-id', 'N2UxNTM3MzItNWNlZC00MDc5LWI3ODUtYWNmZDA2YjE0MWE2-U2VjdGlvbnM=')
+		subscriptionOfferPrompt(flags);
+		electionStub.should.have.callCount(1);
+		document.documentElement.removeAttribute('data-concept-id');
 	});
 
 });
