+++ subscription-offer-prompt/test/us-elections.spec.js
@@ -1,0 +1,101 @@
+/* globals should, sinon */
+import SlidingPopup from 'n-sliding-popup';
+import Superstore from 'superstore';
+import { init } from '../us-elections';
+
+const delay = (ms, value) => new Promise(resolve => setTimeout(resolve.bind(null, value), ms));
+
+describe('"US Election 2016" Subscription Offer Prompt', () => {
+
+	const localStorage = new Superstore('local', 'n-ui.subscription-offer-prompt-us-elections')
+	let flags;
+
+	beforeEach(() => {
+		const fetchStub = sinon.stub(window, 'fetch');
+		fetchStub
+			.withArgs('/country')
+			.returns(Promise.resolve({
+				json: () => Promise.resolve('GBR')
+			}));
+		fetchStub
+			.withArgs('https://next-signup-api.ft.com/offer/62c5c138-aad3-10d7-5aa1-b5a4416f60bd?countryCode=GBR')
+			.returns(Promise.resolve({
+				json: () => Promise.resolve({
+					data: {offer:{id:'62c5c138-aad3-10d7-5aa1-b5a4416f60bd',charges:[{billing_period:'trial',duration_code:'P3M',amount:{symbol:'£',currency:'GBP',value:'50.00'}},{billing_period:'monthly',duration_code:'P1M',amount:{symbol:'£',currency:'GBP',value:'44.00'}}]}}})
+			}));
+
+		// stub out the flag.get()
+		flags = { get: () => true }
+
+		return localStorage.set('last-closed', Date.now() - (1000 * 60 * 60 * 24 * 30));
+	});
+
+	afterEach(() => {
+		window.fetch.restore();
+		flags = null;
+
+		return Promise.all([
+			localStorage.unset('last-closed')
+		]);
+	});
+
+	it('should show prompt if not on a barrier page and hasnt been shown in 30 days', () =>
+		init(flags).then(popup => popup.should.be.an.instanceof(SlidingPopup))
+	);
+
+	it('should have correct attributes', () =>
+		init(flags).then(popup => {
+			popup.el.getAttribute('class').should.include('n-sliding-popup subscription-prompt');
+			popup.el.getAttribute('data-n-component').should.equal('o-sliding-popup');
+			popup.el.getAttribute('data-n-sliding-popup-position').should.equal('bottom left');
+		})
+	);
+
+	it('should have correct html', () =>
+		init(flags).then(popup => {
+			popup.el.innerHTML.should.contain('Your global perspective this US election')
+			popup.el.innerHTML.should.contain('GBP')
+			popup.el.innerHTML.should.contain('£')
+			popup.el.innerHTML.should.contain('50.00')
+		})
+	);
+
+	it('should set onClose to function', () =>
+		init(flags).then(popup => {
+			popup.el.onClose.should.be.a('function')
+		})
+	);
+
+	it('should store date in local storage when closed', () =>
+		init(flags)
+			.then(popup => {
+				popup.el.onClose();
+				return localStorage.get('last-closed');
+			})
+			// give a 1s buffer
+			.then(lastClosed => lastClosed.should.be.closeTo(Date.now(), 1000))
+	);
+
+	// TODO: naughty, but errors for unknown reason - https://circleci.com/gh/Financial-Times/n-ui/2829
+	xit('should ‘pop-up’ after 2 seconds', () =>
+		init(flags)
+			.then(popup => {
+				sinon.spy(popup, 'open');
+				popup.open.should.not.have.been.called;
+				return delay(2050, popup);
+			})
+			.then(popup => popup.open.should.have.callCount(1))
+	);
+
+
+	it('should not show if last shown within 30 days', () => {
+		localStorage.set('last-closed', Date.now() + (1000 * 60 * 60 * 24 * 29));
+		return init(flags).then(popup => should.not.exist(popup));
+	});
+
+	it('should not show if flag usElection2016DiscountSlider is false', () => {
+		flags.get = () => false;
+		return init(flags).then(popup => should.not.exist(popup));
+	});
+
+});
